requirements.txt
fastapi
uvicorn
pandas
requests
sqlalchemy
psycopg2-binary
streamlit
schedule
python-dotenv